html/browsers/browsing-the-web/navigating-across-documents/007.html

<!doctype html>
<title>Link with onclick document.write is aborted by href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="document.getElementById('test').contentWindow.document.write('<script>self.postMessage(&quot;click&quot;, &quot;*&quot;)</script>');" href="href.html">Test</a>
<script>
/*---
specid: navigating-across-documents
steps: [6]
info: |
  Cancel any preexisting but not yet mature attempt to navigate browsingContext, including canceling any instances of the fetch algorithm started by those attempts. If one of those attempts has already created and initialized a new Document object, abort that Document also. (Navigation attempts that have matured already have session history entries, and are therefore handled during the update the session history with the new page algorithm, later.)
description: |
  - This test is operationally equivalent to 005.html with the exeption that:
    - the onclick handler of the anchor element in 007.html selects the iframe in the DOM and sets its `location` to a javascript injection that attempts to post message the string "click.
    - just like 005.html, this tests checks to make sure the href is handled and the onclick's behaviour is aborted before reaching maturity per step 6 of 7.8.1

---*/
var t = async_test();
t.step(function() {
  document.getElementsByTagName("a")[0].click();
});
onmessage = t.step_func(
  function(e) {
    assert_equals(e.data, "href");
    t.done();
  });
</script>
